review comment fix

+++ src/main/java/com/accenture/userservice/service/UserService.java
@@ -1,37 +1,36 @@
 package com.accenture.userservice.service;
 
-import java.util.List;
-
+import com.accenture.userservice.exception.ResourceNotFoundException;
+import com.accenture.userservice.model.User;
+import com.accenture.userservice.repo.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.accenture.userservice.model.User;
-import com.accenture.userservice.repo.UserRepository;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
+import java.util.List;
 
 @Service
 public class UserService {
-	@Autowired
+	
 	private UserRepository userRepository;
-
+	@Autowired
+	public UserService(UserRepository userRepository) {
+		this.userRepository = userRepository;
+	}
 	public User createUser(User user) {
 		return userRepository.save(user);
 	}
-
-	public List<User> createUsers(List<User> users) {
-		return userRepository.saveAll(users);
-	}
-
+	
 	public User getUserById(int id) {
-		return userRepository.findById(id).orElse(null);
+	return	userRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("User Not found with id = " + id));
 	}
-
+	
 	public List<User> getUsers() {
 		return userRepository.findAll();
 	}
-
 	public void deleteUserById(int id) {
+		if (!userRepository.existsById(id)) {
+			throw new ResourceNotFoundException("User Not Found for id "+ id );
+		}
 		userRepository.deleteById(id);		
 	}
